put isValidBirthday on its own file
